Split the molecule and connectivity types in two modules
