+++ src/core/src/sys/molecules.rs
@@ -1,8 +1,7 @@
 // Lumol, an extensible molecular simulation engine
 // Copyright (C) 2015-2016 G. Fraux — BSD license
 
-//! Data about bonds and angles in the system.
-use std::cmp::{min, max};
+//! Data about molecules in the system.
 use std::collections::HashSet;
 use std::iter::IntoIterator;
 use std::ops::Range;
@@ -10,133 +9,8 @@ use std::hash::{Hash, Hasher};
 use std::collections::hash_map::DefaultHasher;
 
 use types::Array2;
-use sys::Particle;
-
-/// A `Bond` between the atoms at indexes `i` and `j`
-///
-/// This structure ensure an unique representation of a `Bond` by enforcing
-/// `i < j`
-#[derive(Clone, Copy, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
-pub struct Bond {
-    i: usize,
-    j: usize,
-}
-
-impl Bond {
-    /// Create a new Bond between the atoms at indexes `first` and `second`
-    pub fn new(first: usize, second: usize) -> Bond {
-        assert!(first != second);
-        let i = min(first, second);
-        let j = max(first, second);
-        Bond{i: i, j: j}
-    }
-
-    /// Get the first particle in the bond
-    #[inline] pub fn i(&self) -> usize {self.i}
-
-    /// Get the second particle in the bond
-    #[inline] pub fn j(&self) -> usize {self.j}
-}
-
-/// An `Angle` formed by the atoms at indexes `i`, `j` and `k`
-///
-/// This structure ensure uniqueness of the `Angle` representation by enforcing
-/// `i < k`
-#[derive(Clone, Copy, Debug, PartialEq, Eq, Hash)]
-pub struct Angle {
-    i: usize,
-    j: usize,
-    k: usize,
-}
-
-impl Angle {
-    /// Create a new Angle between the atoms at indexes `first`, `second` and `third`
-    pub fn new(first: usize, second: usize, third: usize) -> Angle {
-        assert!(first != second);
-        assert!(first != third);
-        assert!(second != third);
-        let i = min(first, third);
-        let k = max(first, third);
-        Angle{i: i, j: second, k: k}
-    }
-
-    /// Get the first particle in the angle
-    #[inline] pub fn i(&self) -> usize {self.i}
-
-    /// Get the second particle in the angle
-    #[inline] pub fn j(&self) -> usize {self.j}
-
-    /// Get the third particle in the angle
-    #[inline] pub fn k(&self) -> usize {self.k}
-}
-
-
-/// A `Dihedral` angle formed by the atoms at indexes `i`, `j`, `k` and `m`
-#[derive(Clone, Copy, Debug, PartialEq, Eq, Hash)]
-pub struct Dihedral {
-    i: usize,
-    j: usize,
-    k: usize,
-    m: usize,
-}
-
-impl Dihedral {
-    /// Create a new Dihedral between the atoms at indexes `first`, `second`,
-    /// `third` and `fourth`
-    pub fn new(first: usize, second: usize, third: usize, fourth: usize) -> Dihedral {
-        assert!(first != second);
-        assert!(second != third);
-        assert!(third != fourth);
-        let (i, j, k, m) = if max(first, second) < max(third, fourth) {
-            (first, second, third, fourth)
-        } else {
-            (fourth, third, second, first)
-        };
-        Dihedral{i: i, j: j, k: k, m: m}
-    }
-
-    /// Get the first particle in the dihedral angle
-    #[inline] pub fn i(&self) -> usize {self.i}
-
-    /// Get the second particle in the dihedral angle
-    #[inline] pub fn j(&self) -> usize {self.j}
-
-    /// Get the third particle in the dihedral angle
-    #[inline] pub fn k(&self) -> usize {self.k}
-
-    /// Get the fourth particle in the dihedral angle
-    #[inline] pub fn m(&self) -> usize {self.m}
-}
-
-/******************************************************************************/
-mod connect {
-    bitflags! {
-        /// The `Connectivity` bitflag encode the topological distance between
-        /// two particles in the molecule, i.e. the number of bonds between the
-        /// particles.
-        pub flags Connectivity: u8 {
-            /// The particles are separated by one bond
-            const CONNECT_12   = 0b0001,
-            /// The particles are separated by two bonds
-            const CONNECT_13   = 0b0010,
-            /// The particles are separated by three bonds
-            const CONNECT_14   = 0b0100,
-            /// The particles are separated by more than three bonds
-            const CONNECT_FAR  = 0b1000,
-        }
-    }
-
-    impl Default for Connectivity {
-        fn default() -> Connectivity {
-            CONNECT_FAR
-        }
-    }
-}
-
-pub use self::connect::Connectivity;
-pub use self::connect::{CONNECT_12, CONNECT_13, CONNECT_14, CONNECT_FAR};
-
-/******************************************************************************/
+use sys::{Particle, Bond, Angle, Dihedral};
+use sys::{Connectivity, CONNECT_12, CONNECT_13, CONNECT_14};
 
 #[derive(Debug, Clone)]
 /// A molecule is the basic building block for a topology. It contains data
@@ -234,14 +108,14 @@ impl Molecule {
                     continue;
                 }
 
-                let angle = if bond1.i == bond2.j {
-                    Angle::new(bond2.i, bond2.j, bond1.j)
-                } else if bond1.j == bond2.i {
-                    Angle::new(bond1.i, bond1.j, bond2.j)
-                } else if bond1.j == bond2.j {
-                    Angle::new(bond1.i, bond1.j, bond2.i)
-                } else if bond1.i == bond2.i {
-                    Angle::new(bond1.j, bond1.i, bond2.j)
+                let angle = if bond1.i() == bond2.j() {
+                    Angle::new(bond2.i(), bond2.j(), bond1.j())
+                } else if bond1.j() == bond2.i() {
+                    Angle::new(bond1.i(), bond1.j(), bond2.j())
+                } else if bond1.j() == bond2.j() {
+                    Angle::new(bond1.i(), bond1.j(), bond2.i())
+                } else if bond1.i() == bond2.i() {
+                    Angle::new(bond1.j(), bond1.i(), bond2.j())
                 } else {
                     // We will not find any dihedral angle from these bonds
                     continue;
@@ -254,14 +128,14 @@ impl Molecule {
                         continue;
                     }
 
-                    let dihedral = if angle.k == bond3.i && angle.j != bond3.j {
-                        Dihedral::new(angle.i, angle.j, angle.k, bond3.j)
-                    } else if angle.k == bond3.j && angle.j != bond3.i {
-                        Dihedral::new(angle.i, angle.j, angle.k, bond3.i)
-                    } else if angle.i == bond3.j && angle.j != bond3.i {
-                        Dihedral::new(bond3.i, angle.i, angle.j, angle.k)
-                    } else if angle.i == bond3.i && angle.j != bond3.j {
-                        Dihedral::new(bond3.j, angle.i, angle.j, angle.k)
+                    let dihedral = if angle.k() == bond3.i() && angle.j() != bond3.j() {
+                        Dihedral::new(angle.i(), angle.j(), angle.k(), bond3.j())
+                    } else if angle.k() == bond3.j() && angle.j() != bond3.i() {
+                        Dihedral::new(angle.i(), angle.j(), angle.k(), bond3.i())
+                    } else if angle.i() == bond3.j() && angle.j() != bond3.i() {
+                        Dihedral::new(bond3.i(), angle.i(), angle.j(), angle.k())
+                    } else if angle.i() == bond3.i() && angle.j() != bond3.j() {
+                        Dihedral::new(bond3.j(), angle.i(), angle.j(), angle.k())
                     } else {
                         // TODO: (angle.k == bond3.i || angle.k == bond3.j)
                         // is an improper dihedral.
@@ -290,18 +164,18 @@ impl Molecule {
         };
 
         for bond in &self.bonds {
-            add_connect_term(bond.i, bond.j, CONNECT_12);
-            add_connect_term(bond.j, bond.i, CONNECT_12);
+            add_connect_term(bond.i(), bond.j(), CONNECT_12);
+            add_connect_term(bond.j(), bond.i(), CONNECT_12);
         }
 
         for angle in &self.angles {
-            add_connect_term(angle.i, angle.k, CONNECT_13);
-            add_connect_term(angle.k, angle.i, CONNECT_13);
+            add_connect_term(angle.i(), angle.k(), CONNECT_13);
+            add_connect_term(angle.k(), angle.i(), CONNECT_13);
         }
 
         for dihedral in &self.dihedrals {
-            add_connect_term(dihedral.i, dihedral.m, CONNECT_14);
-            add_connect_term(dihedral.m, dihedral.i, CONNECT_14);
+            add_connect_term(dihedral.i(), dihedral.m(), CONNECT_14);
+            add_connect_term(dihedral.m(), dihedral.i(), CONNECT_14);
         }
     }
 
@@ -390,19 +264,19 @@ impl Molecule {
         // Remove bonds containing the particle `i`
         let mut new_bonds = HashSet::new();
         for bond in self.bonds() {
-            if bond.i == i || bond.j == i {
+            if bond.i() == i || bond.j() == i {
                 continue;
             }
 
-            let mut new_bond = *bond;
-            if new_bond.i > i {
-                new_bond.i -= 1;
+            let (mut bond_i, mut bond_j) = (bond.i(), bond.j());
+            if bond_i > i {
+                bond_i -= 1;
             }
-            if new_bond.j > i {
-                new_bond.j -= 1;
+            if bond_j > i {
+                bond_j -= 1;
             }
 
-            new_bonds.insert(new_bond);
+            new_bonds.insert(Bond::new(bond_i, bond_j));
         }
 
         self.bonds = new_bonds;
@@ -475,34 +349,11 @@ impl<'a> IntoIterator for &'a Molecule {
     }
 }
 
-/******************************************************************************/
 #[cfg(test)]
 mod test {
     use super::*;
-
-    #[test]
-    fn unicity() {
-        let bond = Bond::new(2, 1);
-        assert_eq!(bond.i, 1);
-        assert_eq!(bond.j, 2);
-
-        let angle = Angle::new(8, 7, 6);
-        assert_eq!(angle.i, 6);
-        assert_eq!(angle.j, 7);
-        assert_eq!(angle.k, 8);
-
-        let dihedral = Dihedral::new(8, 7, 6, 0);
-        assert_eq!(dihedral.i, 0);
-        assert_eq!(dihedral.j, 6);
-        assert_eq!(dihedral.k, 7);
-        assert_eq!(dihedral.m, 8);
-
-        let dihedral = Dihedral::new(0, 7, 6, 8);
-        assert_eq!(dihedral.i, 0);
-        assert_eq!(dihedral.j, 7);
-        assert_eq!(dihedral.k, 6);
-        assert_eq!(dihedral.m, 8);
-    }
+    use sys::{Bond, Angle, Dihedral};
+    use sys::{CONNECT_12, CONNECT_13, CONNECT_14};
 
     #[test]
     fn translate() {
